// src/types/vehicle.ts
export interface VehicleDetails {
  artEndDate: string;
  co2Emissions: number;
  colour: string;
  engineCapacity: number;
  fuelType: string;
  make: string;
  markedForExport: boolean;
  monthOfFirstRegistration: string;
  motStatus: string;
  registrationNumber: string;
  revenueWeight: number;
  taxDueDate: string;
  taxStatus: string;
  typeApproval: string;
  wheelplan: string;
  yearOfManufacture: number;
  euroStatus: string;
  realDrivingEmissions: string;
  dateOfLastV5CIssued: string;
}